requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0

# Database
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
cryptography==41.0.7
bcrypt==4.1.2
PyMySQL==1.1.0
email-validator==2.1.0

# File Parsing
python-docx==1.1.0
python-pptx==0.6.23

# Excel Generation
openpyxl==3.1.2
python-magic==0.4.27

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Development
Flask-DebugToolbar==0.14.1
